src/pyproject.toml
[tool.poetry]
name = "utmb-iis"
version = "0.1.0"
description = "utm builder for asus for windows server"
authors = ["Eduard Kulyk <kulikeduard89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
apscheduler = "^3.10.4"
flask = "^3.0.3"
flask-apscheduler = "^1.13.1"
google-api-python-client = "^2.129.0"
google-auth-oauthlib = "^1.2.0"
google-analytics-data = "^0.18.8"
pandas = "^2.2.2"
pillow = "^10.3.0"
requests = "^2.31.0"
waitress = "^3.0.0"
xlsxwriter = "^3.2.0"
sqlalchemy = {extras = ["mypy"], version = "^2.0.30"}
python-dotenv = "^1.0.1"
pydantic = "^2.7.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.4.2"
ruff = "^0.4.4"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 120
skip = ["wfastcgi.py"]
profile = "black"

[tool.black]
line_length = 120
exclude ='''
    ^wfastcgi\.py
    ^__init__\.py
'''


[tool.mypy]
ignore_missing_imports = true
no_implicit_optional = false
exclude = ["wfastcgi.py","__init__.py"]


[tool.ruff]
exclude = ["wfastcgi.py","__init__.py"]